Added global training/validation functions
1) get_signals_main_parameters.m compute the week1 task1 signal main
parameters
2)color_segmentation_method_... function loads the training or the
validation split images and compute the selected color segmentation
method, storing the output masks in a separate folder, computing the
time and also the validation parameters.
3)signal_detection_method_... function loads the masks generated with
the color segmentation method (the train or validation split computed
maks) and calls a detection method/enhance method to clean/detect
signals from the input maks. The new masks are saved in a separate
folder and the time and validation parameters are also computed.
This last function is useful to call a method that applies morphological
operators to the maks, or to call a method that uses masks to detect
possible signals, as in week2 task1 and 2.